+++ sensorClientRaspberryPi/lib/sensor/ds18b20.py
@@ -91,6 +91,9 @@ class RaspberryPiDS18b20Sensor(_NumberSensor):
             return self.data
 
     def initialize(self) -> bool:
+        if not super().initialize():
+            return False
+
         self.state = 1 - self.triggerState
 
         self._sensor_file = "/sys/bus/w1/devices/" \